run actor bundle tests in sequence
