+++ tests/tool_tests.rs
@@ -7,7 +7,6 @@ use common::{daemon, CommonArgs};
 use predicates::prelude::*;
 
 use crate::common::{create_tmp_config, tool};
-use fickle::fickle;
 
 // Exporting an empty archive should fail but not panic
 #[test]
@@ -25,24 +24,6 @@ fn export_empty_archive() {
         .stderr(predicate::eq(
             "Error: input not recognized as any kind of CAR data (.car, .car.zst, .forest.car)\n",
         ));
-}
-
-// Running `forest-tool state-migration actor-bundle` may not fail.
-#[test]
-#[fickle]
-fn state_migration_actor_bundle_runs() {
-    let temp_dir = tempfile::tempdir().unwrap();
-    let bundle = temp_dir.path().join("bundle.car");
-
-    tool()
-        .arg("state-migration")
-        .arg("actor-bundle")
-        .arg(&bundle)
-        .assert()
-        .success();
-
-    assert!(bundle.exists());
-    assert!(zstd::decode_all(std::fs::File::open(&bundle).unwrap()).is_ok());
 }
 
 #[test]
